Grid gone
Deleted the grid and replaced it with two tables. The first empty table will be used as a placeholder to represent the grid that will be filled in with using java script. The second table will be used as our number selector which in the future will allow us to make inputs into the grid.